✔ Correct option assignment for rejectUnauthorized
Removed the unnecessary default value assignment of the 'rejectUnauthorized' option.

Due to the way the default value was assigned, any attempt to set `rejectUnauthorized` to `false` would just result in the default value being used instead. Now we leave the defaults up to Node, which is probably the better approach anyway. 👍
(`false || true` === `true`)