Add autoEHLO and fix a bug
autoEHLO will allow you to always EHLO to a mail server. This is a configurable option. But, there was a bug where it was continually trying to upgrade the connection because we didn't see that we were in a regular HELO and not a EHLO. This didn't give us the option to upgrade, and this prevents us from sending emails!